+++ lib/sensitive_data_filter/middleware/occurrence.rb
@@ -34,7 +34,7 @@ module SensitiveDataFilter
         @filtered_env_parser.body_params
       end
 
-      def_delegators :@original_env_parser, :request_method, :url, :session
+      def_delegators :@original_env_parser, :request_method, :url, :content_type, :session
 
       def matches_count
         @matches.map { |type, matches| [type, matches.count] }.to_h
@@ -45,6 +45,7 @@ module SensitiveDataFilter
           origin_ip:             origin_ip,
           request_method:        request_method,
           url:                   url,
+          content_type:          content_type,
           filtered_query_params: filtered_query_params,
           filtered_body_params:  filtered_body_params,
           session:               session,